Add files via upload
These files are JSON files that drive the table manager.  One is a template of the actual JSON data structure that needs to be used and the other has comments in it for documentation.  To use table manager you will need to copy and edit TblMgrTemplate.json.  Alternatively you could have a backend process completely or partially generate the JSON structure, but you will still need to understand the template.